+++ SobokSobok/SobokSobok/Presentation/Schedule/Controllers/ScheduleViewController+FSCalendar.swift
@@ -5,4 +5,37 @@
 //  Created by taehy.k on 2022/06/22.
 //
 
-import Foundation
+// MARK: - FSCalendar Setup
+
+import FSCalendar
+
+extension ScheduleViewController {
+    func setCalendar() {
+        calendarView.locale = Locale(identifier: "ko_KR")
+        calendarView.headerHeight = 0
+        calendarView.firstWeekday = 2
+        calendarView.setScope(.week, animated: false)
+    }
+    
+    func setCalendarStyle() {
+        calendarView.appearance.weekdayFont = TypoStyle.body7.font
+        calendarView.appearance.weekdayTextColor = Color.gray600
+        calendarView.appearance.titleFont = UIFont.font(.pretendardRegular, ofSize: 16)
+        calendarView.appearance.titleDefaultColor = Color.black
+        calendarView.appearance.titleTodayColor = Color.black
+        calendarView.appearance.todayColor = .clear
+        calendarView.placeholderType = .none
+    }
+}
+
+// MARK: - FSCalendar Delegate
+
+extension ScheduleViewController: FSCalendarDelegate {
+    func calendar(_ calendar: FSCalendar, boundingRectWillChange bounds: CGRect, animated: Bool) {
+        calendarHeight = bounds.height
+        calendarView.snp.updateConstraints {
+            $0.height.equalTo(calendarHeight)
+        }
+        self.scrollView.layoutIfNeeded()
+    }
+}
